Update the DETECTION structure to match the latest version of darknet.
On 2019-10-25, darknet was updated to add the Gaussian YOLOv3 layer. This had the effect of changing the detection structure in darknet.h that zmMagik uses to interface to darknet. This commit updates zmMagik to account for the new structure. Unfortunately, the darknet version was not updated to reflect this change. For this reason, only builds of darknet from 2019-10-25 or later should be used with zmMagik.